Try setting a longer proxy timeout (#114)
I've investigated this a bit, but it's difficult to confirm whether it
works when SSHing into a dyno. The application still runs at least, but
not sure whether it'll help anything. I'd like to get onto QA to test
it, or possibly just production - I don't think we can really end up in
a worse situation than we're currently in 🙈